Update crime_1.py
The model works as described. The model gets the data as the geo location and the date and time and then using this details the model makes the prediction. This data uses the details and then it works well by taking these details and then makes the prediction as to what are the chances of crime in the region.